third_party/WebKit/LayoutTests/external/wpt/html/rendering/non-replaced-elements/the-page/iframe-marginwidth-marginheight-expected.txt

This is a testharness.js-based test.
FAIL iframe marginwidth and marginheightUncaught Error: assert_equals: Number of attributes on the child document's body expected 0 but got 2
Harness: the test ran to completion.
